requirements.txt
apiclient==1.0.3
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.11.5
cffi==1.11.4
chardet==3.0.4
cssselect==1.0.3
cycler==0.10.0
defusedxml==0.5.0
discovery==0.0.4
Django==2.0.1
feedfinder2==0.0.4
feedparser==5.2.1
google-api-python-client==1.6.4
httplib2==0.10.3
idna==2.6
jieba3k==0.35.1
libpcap==1.0.0b7
Logbook==1.2.1
lxml==4.1.1
matplotlib==2.1.2
mysqlclient==1.3.12
newspaper3k==0.2.6
nltk==3.2.5
numpy==1.14.0
oauth2client==4.1.2
oauthlib==2.0.6
Pillow==5.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
py==1.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyJWT==1.5.3
pyparsing==2.2.0
pyshark==0.3.7.11
python-dateutil==2.6.1
python-social-auth==0.2.21
python3-openid==3.1.0
pytz==2017.3
PyYAML==3.12
requests==2.18.4
requests-file==1.4.3
requests-oauthlib==0.8.0
rsa==3.4.2
scikit-learn==0.19.1
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
tldextract==2.2.0
trollius==1.0.4
uritemplate==3.0.0
urllib3==1.22
wordcloud==1.3.1
